+++ op25/gr-op25_repeater/lib/CMakeLists.txt
@@ -41,7 +41,7 @@ list(APPEND op25_repeater_sources
     ysf_tx_sb_impl.cc
     dstar_tx_sb_impl.cc
     vocoder_impl.cc
-    gardner_costas_cc_impl.cc
+    gardner_cc_impl.cc
     p25_frame_assembler_impl.cc
     frame_assembler_impl.cc
     analog_udp_impl.cc
