+++ op25/gr-op25_repeater/swig/op25_repeater_swig.i
@@ -2,14 +2,14 @@
 
 #define OP25_REPEATER_API
 
-%include "gnuradio.i"			// the common stuff
+%include "gnuradio.i"  // the common stuff
 
 //load generated python docstrings
 %include "op25_repeater_swig_doc.i"
 
 %{
 #include "op25_repeater/vocoder.h"
-#include "op25_repeater/gardner_costas_cc.h"
+#include "op25_repeater/gardner_cc.h"
 #include "op25_repeater/p25_frame_assembler.h"
 #include "op25_repeater/frame_assembler.h"
 #include "op25_repeater/analog_udp.h"
@@ -37,8 +37,8 @@ GR_SWIG_BLOCK_MAGIC2(op25_repeater, dstar_tx_sb);
 %include "op25_repeater/vocoder.h"
 GR_SWIG_BLOCK_MAGIC2(op25_repeater, vocoder);
 
-%include "op25_repeater/gardner_costas_cc.h"
-GR_SWIG_BLOCK_MAGIC2(op25_repeater, gardner_costas_cc);
+%include "op25_repeater/gardner_cc.h"
+GR_SWIG_BLOCK_MAGIC2(op25_repeater, gardner_cc);
 
 %include "op25_repeater/p25_frame_assembler.h"
 GR_SWIG_BLOCK_MAGIC2(op25_repeater, p25_frame_assembler);
